refactor(actions): break up ListDatasets into remote.Client and base
