+++ lib/datasets_test.go
@@ -238,19 +238,21 @@ func TestDatasetRequestsList(t *testing.T) {
 
 	mr, err := testrepo.NewTestRepo()
 	if err != nil {
-		t.Fatalf("error allocating test repo: %s", err.Error())
+		t.Fatalf("error allocating test repo: %s", err)
 		return
 	}
 
 	refs, err := mr.References(0, 30)
 	if err != nil {
-		t.Fatalf("error getting namespace: %s", err.Error())
+		t.Fatalf("error getting namespace: %s", err)
 	}
 
 	node, err := p2p.NewQriNode(mr, config.DefaultP2PForTesting())
 	if err != nil {
-		t.Fatal(err.Error())
+		t.Fatal(err)
 	}
+
+	inst := NewInstanceFromConfigAndNode(config.DefaultConfigForTesting(), node)
 
 	for _, ref := range refs {
 		switch ref.Name {
@@ -285,7 +287,7 @@ func TestDatasetRequestsList(t *testing.T) {
 		// TODO: re-enable {&ListParams{OrderBy: "name", Limit: 30, Offset: 0}, []*repo.DatasetRef{cities, counter, movies}, ""},
 	}
 
-	req := NewDatasetRequests(node, nil)
+	req := NewDatasetRequestsInstance(inst)
 	for _, c := range cases {
 		got := []repo.DatasetRef{}
 		err := req.List(c.p, &got)
